requirements.txt
asgiref==3.3.4
astroid==2.5.6
autopep8==1.5.7
certifi==2021.5.30
chardet==4.0.0
colorama==0.4.4
coreapi==2.3.3
coreschema==0.0.4
Django==3.2.4
djangorestframework==3.12.4
idna==2.10
isort==5.9.1
itypes==1.2.0
Jinja2==3.0.1
lazy-object-proxy==1.6.0
MarkupSafe==2.0.1
mccabe==0.6.1
openapi-codec==1.3.2
pycodestyle==2.7.0
pytz==2021.1
requests==2.25.1
simplejson==3.17.2
sqlparse==0.4.1
toml==0.10.2
uritemplate==3.0.1
urllib3==1.26.6
wrapt==1.12.1
